// backend/src/types.ts
export type Execution = Step[]

export type Step = Persistence | Delete | Animate | Global | Conditional

// --

export type Persistence = Create | Update
// ...
export type Delete = {
  type: 'delete'
  identifier: string
  entity: EntityTypes
}

// --

export type Animate = {
  type: 'animate'
  animationIdentifier: Identifier
  entity: EntityTypes.layer | EntityTypes.object
  entityIdentifier: Identifier
  action: Action
}

export enum Action {
  start = 'start',
  stop = 'stop',
  speedUp = 'speedUp',
  slowDown = 'slowDown',
  reset = 'reset', // TODO: what does this look like for animations? Origin (return-after-loop) from Milestone 1 - Al
}

// --

type Global = Collision | Wait

export type Collision = {
  type: 'collision'
  value: boolean
}

export type Wait = {
  type: 'wait'
  duration: number
}

// --

export type Conditional = {
  type: 'if' | 'loop'
  condition: Condition
  count?: number
  body: Step[]
}

export type OP = '<' | '>' | '<=' | '>=' | '==' | '!='